requirements.txt
altair==4.2.2
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
branca==0.7.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
curl-cffi==0.7.1
datetime==5.5
entrypoints==0.4
folium==0.17.0
gitdb==4.0.11
gitpython==3.1.43
gobnb==0.0.11
idna==3.8
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
narwhals==1.6.3
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.0
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pytest==7.4.3
pytest-cov==4.1.0
pyyaml==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.38.0
streamlit-aggrid==1.0.5
streamlit-folium==0.22.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==5.0.2
xyzservices==2024.9.0
zope-interface==7.0.3